Refactor loop index type in `compute_auto_shard_shape` and enhance core grid initialization in `UnaryDeviceOperation`
### What's changed
- Changed the loop index type from `int` to `size_t` in the `compute_auto_shard_shape` function for better type safety.
- Improved the initialization of `CoreRangeSet` in `UnaryDeviceOperation::compute_output_specs` by using separate `CoreCoord` variables for start and end coordinates.

### Checklist
- [x] Code changes reviewed for correctness
- [x] CI passes for all relevant workflows

+++ ttnn/cpp/ttnn/operations/eltwise/unary/common/unary_op_utils.cpp
@@ -1054,7 +1054,7 @@ std::array<uint32_t, 2> compute_auto_shard_shape(
 
     // Compute "squeezed" 2D shape: all dimensions except width collapsed into height
     uint32_t batch_size = 1;
-    for (int i = 0; i < rank - 2; i++) {
+    for (size_t i = 0; i < rank - 2; i++) {
         batch_size *= logical_shape[i];
     }
